Add file modifications
The main modification is in the calculation of the log-evidence:
There was an issue with the value returned by the function "LogE_SubC_forMCM()". The states of ba that were not observed are supposed to contribute to the numerator as Gamma(1/2) = Sqrt(pi), but they were not counted. This is now corrected. Modification to the values of the logE are very small, unless the considered communities are very large.